Bot::new コンストラクタを追加
