+++ src/lib.rs
@@ -104,6 +104,52 @@ pub struct Bot {
 }
 
 impl Bot {
+    pub fn new(
+        channel_ids: Vec<ChannelId>,
+        debug_channel_id: ChannelId,
+
+        guild_id: GuildId,
+        erogaki_role_id: RoleId,
+        jail_mark_role_id: RoleId,
+        jail_main_role_id: RoleId,
+
+        gemini: ai::GeminiAI,
+
+        commit_hash: Option<String>,
+        commit_date: Option<String>,
+
+        disabled_commands: &[&str]
+    ) -> Self {
+
+        let userdata = DashMap::new();
+        let variables = DashMap::new();
+        let jail_process = Arc::new(DashMap::new());
+        let jail_id = Arc::new(Mutex::new(0));
+        let reply_to_all_mode = Arc::new(Mutex::new(ReplyToAllModeData::blank()));
+        let prefix_commands = prefix_commands(&disabled_commands);
+        let slash_commands = slash_commands(&disabled_commands);
+
+        Self {
+            userdata,
+            jail_process,
+            jail_id,
+            channel_ids,
+            debug_channel_id,
+            guild_id,
+            erogaki_role_id,
+            jail_mark_role_id,
+            jail_main_role_id,
+            commit_hash,
+            commit_date,
+            variables,
+            reply_to_all_mode,
+            gemini,
+            prefix_commands,
+            slash_commands,
+        }
+    }
+
+
     pub fn get_user_room_pointer(&self, user_id: &UserId) -> ChannelId {
         self.userdata
             .entry(*user_id)
